Flexibly get tensor shapes - Different versions of tensorflow/keras were returning `tuple` or `tf.TensorShape`.     - Create new function to handle both     - refactor utils to create this function and the current model loading util functions.
