+++ dosma/models/__init__.py
@@ -1,12 +1,13 @@
-from dosma.models import oaiunet2d, stanford_qdess, stanford_qdess_bone, util
+from dosma.models import model_loading_util, oaiunet2d, stanford_qdess, stanford_qdess_bone
 from dosma.models.oaiunet2d import *  # noqa
 from dosma.models.stanford_qdess import *  # noqa: F401, F403
 from dosma.models.stanford_qdess_bone import *  # noqa: F401, F403
 from dosma.models.stanford_cube_bone import *  # noqa: F401, F403
+from dosma.models.model_loading_util import *  # noqa
 from dosma.models.util import *  # noqa
 
 __all__ = []
-__all__.extend(util.__all__)
+__all__.extend(model_loading_util.__all__)
 __all__.extend(oaiunet2d.__all__)
 __all__.extend(stanford_qdess.__all__)
 __all__.extend(stanford_qdess_bone.__all__)
