+++ dosma/models/stanford_qdess.py
@@ -27,6 +27,7 @@ except ImportError:  # pragma: no-cover
 from dosma.core.med_volume import MedicalVolume
 from dosma.core.orientation import SAGITTAL
 from dosma.models.seg_model import KerasSegModel, whiten_volume
+from dosma.models.util import get_tensor_shape_as_list
 
 __all__ = ["StanfordQDessUNet2D"]
 
@@ -108,7 +109,8 @@ class StanfordQDessUNet2D(KerasSegModel):
             if depth_cnt < depth - 1:
 
                 # If size of input is odd, only do a 3x3 max pool
-                xres = conv.shape.as_list()[1]
+                shape_ = get_tensor_shape_as_list(conv)
+                xres = shape_[1]
                 if xres % 2 == 0:
                     pooling_size = (2, 2)
                 elif xres % 2 == 1:
@@ -119,7 +121,7 @@ class StanfordQDessUNet2D(KerasSegModel):
         # step up convolutional layers
         for depth_cnt in range(depth - 2, -1, -1):
 
-            deconv_shape = conv_ptr[depth_cnt].shape.as_list()
+            deconv_shape = get_tensor_shape_as_list(conv_ptr[depth_cnt])
             deconv_shape[0] = None
 
             # If size of input is odd, then do a 3x3 deconv
